Made both fibs use long data types

+++ fib.c
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 #include <assert.h>
 
-int fib (int x)
+long fib (long x)
 {
 	if(x <= 2) return 1;
 	else return (fib(x-2) + fib(x-1));
@@ -14,6 +14,6 @@ int fib (int x)
 int main(int argc, char *argv[]) {
 	int x = atoi(argv[1]);
 	printf("arg is: %d\n", x);
-	printf("fib(%d) = %d\n",x,fib(x));
+	printf("fib(%d) = %ld\n",x,fib((long)x));
 	return 0;
 }
